Move code into a new server module called 'error-reporting-module'
