+++ settings.gradle
@@ -20,5 +20,6 @@ include 'smoke-testing'
 include 'spitfire-server:database'
 include 'spitfire-server:database-test-support'
 include 'spitfire-server:dropwizard-server'
+include 'spitfire-server:error-reporting-module'
 include 'spitfire-server:lobby-module'
 include 'spitfire-server:maps-module'
