+++ spitfire-server/dropwizard-server/build.gradle
@@ -55,6 +55,7 @@ dependencies {
     implementation project(':lib:java-extras')
     implementation project(':lib:websocket-client')
     implementation project(':lib:websocket-server')
+    implementation project(':spitfire-server:error-reporting-module')
     implementation project(':spitfire-server:lobby-module')
     implementation project(':spitfire-server:maps-module')
     testImplementation "com.github.database-rider:rider-junit5:$databaseRiderVersion"
