+++ spitfire-server/error-reporting-module/src/main/java/org/triplea/modules/error/reporting/CreateIssueParams.java
