+++ spitfire-server/error-reporting-module/src/test/java/org/triplea/modules/error/reporting/CreateIssueStrategyTest.java
@@ -12,12 +12,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.triplea.db.dao.error.reporting.ErrorReportingDao;
-import org.triplea.db.dao.error.reporting.InsertHistoryRecordParams;
 import org.triplea.http.client.error.report.ErrorReportRequest;
 import org.triplea.http.client.error.report.ErrorReportResponse;
 import org.triplea.http.client.github.CreateIssueResponse;
 import org.triplea.http.client.github.GithubApiClient;
+import org.triplea.modules.error.reporting.db.ErrorReportingDao;
+import org.triplea.modules.error.reporting.db.InsertHistoryRecordParams;
 
 @ExtendWith(MockitoExtension.class)
 class CreateIssueStrategyTest {
