+++ spitfire-server/error-reporting-module/src/test/resources/db-cleanup.sql
@@ -1,0 +1,3 @@
+-- Deletes table data in proper order
+
+delete from error_report_history;
