+++ spitfire-server/error-reporting-module/src/test/resources/dbunit.yml
@@ -1,0 +1,6 @@
+caseInsensitiveStrategy: !!com.github.database.rider.core.api.configuration.Orthography 'LOWERCASE'
+connectionConfig:
+  driver: "org.postgresql.Driver"
+  url: "jdbc:postgresql://localhost:5432/lobby_db"
+  user: "lobby_user"
+  password: "lobby"
