+++ spitfire-server/lobby-module/build.gradle
@@ -13,12 +13,10 @@ dependencies {
     implementation "org.jdbi:jdbi3-sqlobject:$jdbiVersion"
     implementation project(':game-app:domain-data')
     implementation project(':http-clients:feign-common')
-    implementation project(':http-clients:github-client')
     implementation project(':http-clients:lobby-client')
     implementation project(':lib:java-extras')
     implementation project(':lib:websocket-client')
     implementation project(':lib:websocket-server')
-    implementation project(':spitfire-server:database')
     testImplementation "com.github.database-rider:rider-junit5:$databaseRiderVersion"
     testImplementation "com.sun.mail:jakarta.mail:$jakartaMailVersion"
     testImplementation "io.dropwizard:dropwizard-testing:$dropwizardVersion"
