+++ spitfire-server/lobby-module/src/test/resources/db-cleanup.sql
@@ -4,7 +4,6 @@ delete from access_log;
 delete from bad_word;
 delete from banned_user;
 delete from banned_username;
-delete from error_report_history;
 delete from game_chat_history;
 delete from lobby_game;
 delete from game_hosting_api_key;
